+++ orm/typedef/IDbMapping.ts
@@ -56,4 +56,8 @@ export interface IDbMapping {
      */
     relations: Array<IRelationDescriptor<unknown>>;
 
+    /**
+     * When using custom ID, this method will check if the Entity is new.
+     */
+    isNew?: () => boolean;
 }