+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Noteworthy
 
   - Fixed `orphanRemoval` in combination with `identifies` option.
+  - Added `isNew` option for saving attributes with a custom id.